+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - backend: expose the reason why a thing / list is visible to a user
+- backend: add `orderBy` and `orderDirection` to `/api/things`, allows to display directly shared things first
 - add defaultSharingState to profile, creates things and lists with that sharing setting
 - development: a nix-based development environment is available
   with the `dev` flake output.